Cleans up tests
