+++ Projects/Server.Tests/Network/Packets/Utilities/PacketTestUtilities.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Net.Sockets;
 using Server.Network;
 
 namespace Server.Tests.Network
@@ -10,6 +9,6 @@ namespace Server.Tests.Network
             p.Compile(false, out var length).AsSpan(0, length);
 
         public static NetState CreateTestNetState() =>
-            new NetState(new Socket(SocketType.Stream, ProtocolType.Tcp));
+            new NetState(new MockSocket());
     }
 }
