+++ Projects/Server/Network/ISocket.cs
@@ -1,0 +1,35 @@
+/*************************************************************************
+ * ModernUO                                                              *
+ * Copyright 2019-2020 - ModernUO Development Team                       *
+ * Email: hi@modernuo.com                                                *
+ * File: ISocket.cs                                                      *
+ *                                                                       *
+ * This program is free software: you can redistribute it and/or modify  *
+ * it under the terms of the GNU General Public License as published by  *
+ * the Free Software Foundation, either version 3 of the License, or     *
+ * (at your option) any later version.                                   *
+ *                                                                       *
+ * You should have received a copy of the GNU General Public License     *
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
+ *************************************************************************/
+
+using System;
+using System.Net;
+using System.Net.Sockets;
+using System.Threading.Tasks;
+
+namespace Server.Network
+{
+    public interface ISocket
+    {
+        public EndPoint LocalEndPoint { get; }
+
+        public EndPoint RemoteEndPoint { get; }
+
+        public Task<int> SendAsync(ArraySegment<byte>[] buffer, SocketFlags flags);
+
+        public Task<int> ReceiveAsync(ArraySegment<byte>[] buffer, SocketFlags flags);
+
+        public void Shutdown(SocketShutdown how);
+    }
+}
