+++ Projects/Server/Network/NetState/NetState.cs
@@ -75,7 +75,7 @@ namespace Server.Network
             Timer.DelayCall(checkAliveDuration, checkAliveDuration, CheckAllAlive);
         }
 
-        public NetState(Socket connection)
+        public NetState(ISocket connection)
         {
             m_Running = false;
             Connection = connection;
@@ -141,7 +141,7 @@ namespace Server.Network
 
         public Pipe<byte> SendPipe { get; private set; }
 
-        public Socket Connection { get; private set; }
+        public ISocket Connection { get; private set; }
 
         public bool CompressionEnabled { get; set; }
