+++ Projects/Server/Network/NetworkSocket.cs
@@ -1,0 +1,39 @@
+/*************************************************************************
+ * ModernUO                                                              *
+ * Copyright 2019-2020 - ModernUO Development Team                       *
+ * Email: hi@modernuo.com                                                *
+ * File: NetworkSocket.cs                                                *
+ *                                                                       *
+ * This program is free software: you can redistribute it and/or modify  *
+ * it under the terms of the GNU General Public License as published by  *
+ * the Free Software Foundation, either version 3 of the License, or     *
+ * (at your option) any later version.                                   *
+ *                                                                       *
+ * You should have received a copy of the GNU General Public License     *
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
+ *************************************************************************/
+
+using System;
+using System.Net;
+using System.Net.Sockets;
+using System.Threading.Tasks;
+
+namespace Server.Network
+{
+    public class NetworkSocket : ISocket
+    {
+        public Socket Connection { get; }
+        public EndPoint LocalEndPoint => Connection.LocalEndPoint;
+        public EndPoint RemoteEndPoint => Connection.RemoteEndPoint;
+
+        public NetworkSocket(Socket connection) => Connection = connection;
+
+        public Task<int> SendAsync(ArraySegment<byte>[] buffer, SocketFlags flags) =>
+            Connection.SendAsync(buffer, flags);
+
+        public Task<int> ReceiveAsync(ArraySegment<byte>[] buffer, SocketFlags flags) =>
+            Connection.ReceiveAsync(buffer, flags);
+
+        public void Shutdown(SocketShutdown how) => Connection.Shutdown(how);
+    }
+}
